requirements.txt
python-qbittorrent==0.4.3
